Add files via upload
I tried to sort out the decision table. It took me a while to figure out the different diagnostic markers. I changed the "Flow Cytometrie" to a White Blood Cell count test because it is easier to differentiate a diagnosis. The blood test type columns are Antigen and Rh Protein.  Pregnancy is indicated by figures, they are explained in the annotation.